이모티콘 보관함 UI 개선

+++ EmojiPlayground/Sources/UI/EmoticonStorageDetail/EmoticonStorageDetailView.swift
@@ -14,13 +14,15 @@ struct EmoticonStorageDetailView: View {
     let groupName: String
     
     var body: some View {
-        List {
-            LazyVGrid(columns: Array(repeating: GridItem(), count: 3)) {
+        ScrollView {
+            LazyVGrid(columns: Array(repeating: GridItem(spacing: 16), count: 3), spacing: 16) {
                 ForEach(storage.groupImages(groupName: groupName)) { image in
                     EmoticonView(image: image)
                 }
             }
+            .padding()
         }
+        .background(Color.systemGray6)
         .navigationTitle(groupName)
         .navigationBarTitleDisplayMode(.inline)
     }
@@ -32,14 +34,14 @@ struct EmoticonView: View {
     @State private var presentActionAlert: Bool = false
     @State private var groupAlert: Bool = false
     @State private var newGroupName: String = ""
+    @State private var presentDeleteAlert: Bool = false
     
     let image: MessageImage
     
     var body: some View {
         WebImage(url: image.asset.fileURL)
             .resizable()
-            .customLoopCount(4)
-            .scaledToFit()
+            .aspectRatio(1, contentMode: .fit)
             .onTapGesture { /* SCROLLABLE WITH LONG PRESS GESTURE */ }
             .onLongPressGesture {
                 presentActionAlert = true
@@ -53,11 +55,17 @@ struct EmoticonView: View {
                     groupAlert.toggle()
                 }
                 
-                Button("메시지 삭제", role: .destructive) {
-                    storage.delete(image: image)
+                Button("삭제", role: .destructive) {
+                    presentDeleteAlert.toggle()
                 }
                 
                 Button("취소", role: .cancel) { }
+            }
+            .alert("삭제하시겠습니까?", isPresented: $presentDeleteAlert) {
+                Button("삭제", role: .destructive) { storage.delete(image: image) }
+                Button("취소", role: .cancel) { }
+            } message: {
+                Text("삭제된 파일은 복구할 수 없습니다.")
             }
             .sheet(isPresented: $groupAlert) {
                 NavigationView {
